EntregaTerceraPracticaIntegradora

+++ EntregaTerceraPracticaIntegradora/src/controllers/products.controller.js
@@ -1,27 +1,34 @@
+import {
+  getProduct as getProductService,
+  getProductById as getProductByIdService,
+  deleteProduct as deleteProductService,
+  createProduct as createProductService,
+  updateProduct as updateProductService,
+} from "../services/products.service.js";
+import { generateProduct } from "../utils.js";
 import CustomError from "../middlewares/errors/CustomError.js";
 import EErrors from "../middlewares/errors/enums.js";
-import logger from "../utils/logger.js";
 
 export const getProduct = async (req, res) => {
   try {
     const products = await getProductService();
-    res.status(200).send({ status: "success", payload: products });
+    return res.status(200).send({ status: "success", payload: products });
   } catch (error) {
-    logger.error(error);
-    res.status(500).send({ status: "error", error: error.message });
+    return res.send({ status: "error", error: error });
   }
 };
+
 export const getProductById = async (req, res) => {
   try {
     const { pid } = req.params;
     const product = await getProductByIdService(pid);
     if (!product) {
-      res.status(404).send({ status: "error", message: "Product not found" });
-    } else {
-      res.send({ status: "success", payload: product });
+      return res
+        .status(404)
+        .send({ status: "error", message: "Product not found" });
     }
+    res.send({ status: "success", payload: product });
   } catch (error) {
-    logger.error(error);
     res.status(500).send({ error: error.message });
   }
 };
@@ -29,12 +36,18 @@ export const getProductById = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   try {
     const { pid } = req.params;
+    const product = await getProductByIdService(pid);
+    if (product.owner != req.user.email && req.user.role != "admin") {
+      return res.status(403).send({
+        status: "error",
+        message: "Forbidden. You do not own this product.",
+      });
+    }
     const result = await deleteProductService(pid);
-    const io = req.app.get("socketio");
-    io.emit("showProducts", { products: await getProductService() });
+    const io = req.app.get("socketio"); //esto no sé en qué capa debería ir.
+    io.emit("showProducts", { products: await getProductService() }); //esto no sé en qué capa debería ir.
     res.status(201).send({ status: "success", payload: result });
   } catch (error) {
-    logger.error(error);
     res.status(500).send({ status: "error", message: error.message });
   }
 };
@@ -50,18 +63,36 @@ export const createProduct = async (req, res) => {
       category,
       stock,
       status,
+      owner,
     } = req.body;
+    const io = req.app.get("socketio");
 
-    if (!title || !price || title.trim() === "" || price <= 0) {
+    if (
+      !title ||
+      !description ||
+      !price ||
+      !code ||
+      !category ||
+      stock === null ||
+      stock === undefined ||
+      stock === ""
+    ) {
       throw CustomError.createError({
         name: "ProductError",
-        cause: "Title and price are required fields with valid values",
+        cause:
+          "Invalid data types: title (string), description (string), price (number), code (string), category (string) and stock (number) are required",
         message: "Error trying to create product",
         code: EErrors.INVALID_TYPE_ERROR,
       });
     }
-
-    const io = req.app.get("socketio");
+    if (owner != "admin" && owner != req.user.email && owner != null) {
+      throw CustomError.createError({
+        name: "ProductError",
+        cause: 'owner should be your email address (default "admin")',
+        message: "Error trying to create product",
+        code: EErrors.INVALID_TYPE_ERROR,
+      });
+    }
     const product = {
       title,
       description,
@@ -71,25 +102,21 @@ export const createProduct = async (req, res) => {
       category,
       stock,
       status,
+      owner,
     };
-
     const result = await createProductService(product);
-
     if (!result) {
-      res
+      return res
         .status(400)
         .send({ status: "error", message: "product already exists" });
-    } else {
-      const products = await getProductService();
-      io.emit("showProducts", { products: products });
-      res.status(201).send({ status: "success", payload: result });
     }
+    const products = await getProductService();
+    io.emit("showProducts", { products: products });
+    res.status(201).send({ status: "success", payload: result });
   } catch (error) {
-    logger.error(error);
-    res.status(500).send({ status: "error", message: error.message });
+    res.status(400).send({ status: "error", message: error.cause });
   }
 };
-
 export const updateProduct = async (req, res) => {
   try {
     const {
@@ -104,6 +131,7 @@ export const updateProduct = async (req, res) => {
     } = req.body;
     const { pid } = req.params;
     const io = req.app.get("socketio");
+
     const product = {
       title,
       description,
@@ -119,7 +147,20 @@ export const updateProduct = async (req, res) => {
     io.emit("showProducts", { products: products });
     res.status(201).send({ status: "success", payload: result });
   } catch (error) {
-    logger.error(error);
     res.status(500).send({ status: "error", message: error.message });
   }
+};
+
+export const mockingProducts = async (req, res) => {
+  let products = [];
+
+  for (let i = 0; i < 100; i++) {
+    products.push(generateProduct());
+  }
+
+  res.send({
+    status: "ok",
+    counter: products.length,
+    data: products,
+  });
 };
